scripts/requirements.txt
feedparser>=6.0
youtube-transcript-api>=1.0.0
pyyaml>=6.0
requests>=2.31
beautifulsoup4>=4.12
